requirements.txt
fastapi
uvicorn
tensorflow-cpu
numpy
pandas
scikit-learn
joblib
